feat(workflow): add IAM definitions generation workflow (#107)
* feat(workflow): add IAM definitions generation workflow

Introduces a new GitHub Actions workflow for generating IAM definitions. This workflow includes scheduled execution and can be triggered manually. It updates IAM action and managed policy definitions from AWS, enhancing automation and ensuring up-to-date configurations.

* feat(ci): add workflow to create issues from Dependabot alerts

This workflow runs daily to check for open Dependabot alerts and creates GitHub issues for any that do not already have an issue. It provides a structured way to manage security vulnerabilities in dependencies.